feat: Add clean record and dock related traits (#550)
* feat: Add dock summary and clean record traits

This creates a new dock type required feature. This requires the status is fetched first initially.
The dock type is added to the cache since it is not expected to change once discovered.

The clean record trait depends on the clean summary and fetches the last clean record
from the summary.

Tests are added for clean record and dock summary. The dock summary tests also exercise
the feature discovery and underlying modes.

* fix: Remove DockSummary and make dust collection mode optional based on dock type