chore: refactor v1 rpc channels (#609)
* chore: refactor v1 rpc channels

This merges the two packages v1_channel and v1_rpc_channel since they are very closely related. The differences between the RPC types are now handled with `RpcStrategy` (encoding, decoding, which channel, health checking, etc). The "PayloadEncodedV1RpcChannel" is now `_send_rpc` which accepts the rpc strategy.

The `V1RpcChannel` interface is moved to the `v1_protocol` module and now has a single implementation that handles everyting (the `PickFirstAvailable`  logic as well as the response parsing).

Overall the code is now less generalized, but is probably easier to understand since its all in a single place. Notably, all the rpc code was already tested via the v1_channel interface.

* chore: fix v1 channel typing and  improve readability

* chore: remove unnecessary pydoc on private members

* chore: remove unnecessary pydoc to make the code more compact

* chore: Improve doc string readability and grammar

* chore: remove unnecessary docstrings

* fix: remove python 3.11 incompatibility