feat: Add diagnostics library for tracking stats/counters
Add a generic library that can be used to track counters, or elapsed time (e.g. how long it takes on average to connect to mqtt).  This is a copy of https://github.com/allenporter/python-google-nest-sdm/blob/main/google_nest_sdm/diagnostics.py

This is an initial pass to add a few initial example metrics for MQTT, but we can add more as we need fine grained details in diagnostics.