chore: add socket based tests for the new APIs (#677)
* chore: add socket based tests for the new APIs

This adds socket based tests for the new API so we can have tests that exercise the exact byte streams that are sent and received, similar to the old API. These currently are testing the lower level "session" and "channel" just to get the request infrastructure ready to be used in a real end to end device manaver tests.

This violates our rule that tests align with a module, but this is an exception given its for end to end tests.

* chore: Fix local session