fix: Handle AppInitStatus with omitted new_feature_info_str (#688)
* fix: Handle AppInitStatus with omitted new_feature_info_str

The approach taken here is:
(1) add an example of this response type to the json examples to exercise the decoder. (This succeeds already, but is just used to record the payload)
(2) Add a test for passing the decoded data to the dataclass.

In the future this can be improved with trait tests that use these payloads, but this is a way to do it incrementally for now.

* chore: Update default value for new feature string to empty string

* chore: Add additional test coverage for default string value