fix: Reduce log level of decode errors (#691)
* fix: Reduce log level of decode errors

We need to resolve the source of decode errors, however we shouldn't spam the logs since they may not be actionable for users. We should instead focus on fixing symptoms like things broken like RPCs that don't contain a response (e.g. timeouts)

* chore: update tests

* fix: Reduce log level of decode errors

We need to resolve the source of decode errors, however we shouldn't spam the logs since they may not be actionable for users. We should instead focus on fixing symptoms like things broken like RPCs that don't contain a response (e.g. timeouts)

* chore: update tests

* chore: fix tests