chore: Restructure the channel modules
This provides a separation between the lower level protocol
specific channels (MQTT vs Local) and the Application/RPC
level command handling channel modules. This updates the
pydoc to clarify the responsibilities between each layer.

The motivation is from noticing that b01 introduced a number of
new channels, protocols, traits, and it seemed worth separating
for clarity as we want to add new developers to the code base.