chore: Add a01 and b01 q7 byte level tests (#724)
This exercises the byte level behaviors of the last two protocol types. After adding this, we have at least some basic encoding tests of all devices to avoid accidental regressions.

The response builder in the device manager test has been refactored to betters support the v1 , a01, and b01 message building.
This updates the two test devices to have local keys that are the same instead of different per device to make the response building simpler in tests.